[FIX] MergePlugin: move merge-related getter
Task: 4908195
