+++ src/plugins/core/cell.ts
@@ -32,7 +32,7 @@ import { recomputeZones } from "../../helpers/recompute_zones";
 import { Format } from "../../types/format";
 import { Locale } from "../../types/locale";
 import { Style, UpdateCellData, Zone } from "../../types/misc";
-import { Range, RangePart } from "../../types/range";
+import { Range } from "../../types/range";
 import { ExcelWorkbookData, WorkbookData } from "../../types/workbook_data";
 import { SquishedContent, Squisher } from "./squisher";
 import { Unsquisher } from "./unsquisher";
@@ -52,7 +52,6 @@ interface CoreState {
  */
 export class CellPlugin extends CorePlugin<CoreState> implements CoreState {
   static getters = [
-    "zoneToXC",
     "getCells",
     "getTranslatedCellFormula",
     "getCellById",
@@ -427,52 +426,6 @@ export class CellPlugin extends CorePlugin<CoreState> implements CoreState {
       targetSheetId,
       adaptedDependencies
     ).toFormulaString(this.getters);
-  }
-
-  /**
-   * Converts a zone to a XC coordinate system
-   *
-   * The conversion also treats merges as one single cell
-   *
-   * Examples:
-   * {top:0,left:0,right:0,bottom:0} ==> A1
-   * {top:0,left:0,right:1,bottom:1} ==> A1:B2
-   *
-   * if A1:B2 is a merge:
-   * {top:0,left:0,right:1,bottom:1} ==> A1
-   * {top:1,left:0,right:1,bottom:2} ==> A1:B3
-   *
-   * if A1:B2 and A4:B5 are merges:
-   * {top:1,left:0,right:1,bottom:3} ==> A1:A5
-   */
-  zoneToXC(
-    sheetId: UID,
-    zone: Zone,
-    fixedParts: RangePart[] = [{ colFixed: false, rowFixed: false }]
-  ): string {
-    zone = this.getters.expandZone(sheetId, zone);
-    const topLeft = toXC(zone.left, zone.top, fixedParts[0]);
-    const botRight = toXC(
-      zone.right,
-      zone.bottom,
-      fixedParts.length > 1 ? fixedParts[1] : fixedParts[0]
-    );
-    const cellTopLeft = this.getters.getMainCellPosition({
-      sheetId,
-      col: zone.left,
-      row: zone.top,
-    });
-    const cellBotRight = this.getters.getMainCellPosition({
-      sheetId,
-      col: zone.right,
-      row: zone.bottom,
-    });
-    const sameCell = cellTopLeft.col === cellBotRight.col && cellTopLeft.row === cellBotRight.row;
-    if (topLeft !== botRight && !sameCell) {
-      return topLeft + ":" + botRight;
-    }
-
-    return topLeft;
   }
 
   /**
